Testing +++++++++ Currency+

+++ app/Modules/Setting/Entity/Parser.php
@@ -5,7 +5,7 @@ namespace App\Modules\Setting\Entity;
 
 class Parser extends AbstractSetting
 {
-    public int $parser_coefficient = 0;
+    public float $parser_coefficient = 0;
     public int $parser_delivery = 0;
     public int $cost_weight = 0;
     public int $cost_weight_fragile = 0;
